requirements.txt
Django>=6.0,<7.0
PyMuPDF>=1.26,<2.0
openai>=2.0,<3.0
google-generativeai>=0.8,<1.0
Pillow>=12.0,<13.0
python-dotenv>=1.0,<2.0
markdown>=3.10,<4.0
